+++ CHANGELOG.md
@@ -118,6 +118,7 @@
 - New #1013: Add `StringableStream` class to cast binary column values to `string` using `(string) $value` (@Tigrov)
 - Chg #1014: Replace `getEscapingReplacements()`/`setEscapingReplacements()` methods with `escape` constructor parameter 
   in `Like` condition (@vjik)
+- New #1015: Add mode parameter to `Like` condition (@vjik)
 
 ## 1.3.0 March 21, 2024